+++ src/main/resources/META-INF/rewrite/jackson-2-3.yml
@@ -429,16 +429,6 @@ recipeList:
   - org.openrewrite.java.ChangeMethodName:
       methodPattern: com.fasterxml.jackson.databind.JsonNode asText(..)
       newMethodName: asString
-  - org.openrewrite.java.jackson.JsonNodeFieldIteratorMigration
-  - org.openrewrite.java.ChangeMethodName:
-      methodPattern: com.fasterxml.jackson.databind.JsonNode elements()
-      newMethodName: values
-  - org.openrewrite.java.ChangeMethodName:
-      methodPattern: com.fasterxml.jackson.databind.JsonNode fields()
-      newMethodName: properties
-  - org.openrewrite.java.ChangeMethodName:
-      methodPattern: com.fasterxml.jackson.databind.JsonNode fieldNames()
-      newMethodName: propertyNames
   - org.openrewrite.java.ChangeMethodName:
       methodPattern: com.fasterxml.jackson.databind.JsonNode findValuesAsText(..)
       newMethodName: findValuesAsString
@@ -454,6 +444,7 @@ recipeList:
   - org.openrewrite.java.ChangeMethodName:
       methodPattern: com.fasterxml.jackson.databind.JsonNode with(..)
       newMethodName: withObject
+  - org.openrewrite.java.jackson.Jackson3JsonNodeFieldIterators
 
 ---
 type: specs.openrewrite.org/v1beta/recipe
